+++ test/integration/experimental-warning/next.config.js
@@ -1,0 +1,2 @@
+// prettier-ignore
+module.exports = { /* replaceme */ }
