+++ test/integration/experimental-warning/pages/index.js
@@ -1,0 +1,3 @@
+export default function Page() {
+  return <h1>Hi</h1>
+}
